backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pip==25.0.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3